BGM再生の処理をMVPに書き換える。
